requirements.txt
ultralytics
roboflow==1.2.11
onnx==1.19.1
onnxruntime-gpu==1.23.2
tensorrt==10.0.1
tensorrt-cu12==10.14.1.48.post1
pycuda==2025.1.2
numpy==2.0.2
opencv-python==4.12.0.88
PyYAML==6.0.3
Pillow==11.3.0
albumentations==2.0.8